requirements.txt
discord.py==0.16.12
profanity==1.1
